Refactor intl dirs  structure
