+++ app/src/intl/locales.ts
@@ -146,7 +146,7 @@ export const loadOne = (locale: string): ILocale =>
 			(item) => item.locale.includes(locale)
 		)[0];
 
-		res.messages = require(`./${res.file}`);
+		res.messages = require(`./translations/${res.file}`);
 	}
 	catch
 	{
