+++ app/src/intl/translations/cn.json
