+++ app/src/intl/translations/fr.json
