+++ app/src/intl/translations/kk.json
