+++ app/src/intl/translations/nb.json
