+++ app/src/intl/translations/pl.json
